# src/app/api/docs/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "ProjectFlow API",
    "version": "1.0.0",
    "description": "REST API for ProjectFlow - Smart Project Collaboration Platform",
    "contact": {
      "name": "API Support",
      "email": "support@projectflow.com",
      "url": "https://projectflow.com/support"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "https://api.projectflow.com/v1",
      "description": "Production server"
    },
    {
      "url": "https://staging-api.projectflow.com/v1",
      "description": "Staging server"
    },
    {
      "url": "http://localhost:3000/api",
      "description": "Development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "auth-token"
      }
    },
...
  },
...
  "paths": {
    "/auth/login": {
      "post": {
        "tags": ["Authentication"],
        "summary": "User login",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["email", "password"],
                "properties": {
                  "email": { "type": "string", "format": "email" },
                  "password": { "type": "string" },
                  "rememberMe": { "type": "boolean", "default": false }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Login successful" },
          "401": { "description": "Invalid credentials" }
        }
      }
    },
...
  }
}